requirements.txt
flask==2.3.3
pillow==10.0.0
piexif==1.1.3
gunicorn 